backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
aiohttp==3.9.5
SQLAlchemy==2.0.31
python-telegram-bot==13.15
google-api-python-client==2.136.0
notion-client==2.1.1
python-dotenv==1.0.1
pydantic==2.7.4
